Add Event* types for functions with ignored result
When requesting types, either none or all of the generic types must be
supplied, but there is no middle-ground. Because there is often a use
case for objects of type Function where the result is ignored, a
matching family of Event types has been added that types the parameters
but not the return type.