Add cast method to types for foreign objects
Many of the runtime methods assumed that they were always given Grace
objects as parameters, but this was not necessarily true. Now all types
have cast() method which works as assert() but, if given a JavaScript
object, will return a Grace object that wraps the signatures in that
type over the actual parameter. A number of cases where cast was
unnecessary but a change was required to handle JavaScript objects has
also been added.