Report method return pattern on failed assertion
A failed assertion on a method return pattern reported the request site
as the original cause of the error. The pattern on the method is now the
subject of the first line in the stack trace.

The AST signatures were not being correctly stringified after the change
to their properties. This has also been resolved.