Ban self-dependent types, report missing names
Failure to satisfy a type reported the value and the type with the
exception, but did not mention the missing method name. Now the type
matching raises a more specialised error that reports the first
signature name that was not a method in the object.

In order to facilitate this change, all self-dependent types will now
raise an Invalid Type error, as expected. The type proxy logic now
facilitates reacting to type resolution, as well as handling
dependencies between proxies, and self-dependent types are detected by
checking if any types have not been fully resolved once the type
hoisting phase is over. Producing an 'and' pattern that contains at
least one unresolved proxy that produces a new proxy that will resolve
once both of its dependencies do, at which point, if the two
dependencies turn out to be types, it will receive an assert() method
that reports missing method names as well.