Convert primitive method results to Grace objects
The lookup procedure pulls out methods from built-in Grace objects when
detecting JavaScript primitives as receivers of method requests, which
comes with the assumption that the outcome of the method was also a
Grace object. This was not necessarily true, particularly for lists, so
the outcome is now converted from a primitive.