Revert ambiguity, resolve to nearest definition
The ambiguity rules for lookups made inheriting from a dialect
practically impossible to use. As such, the ambiguity error has been
removed, and the lookup will always resolve to the first definition that
it finds. A test for a bug that existed in the old lookup system where
the most distant name was resolved instead of the closes has been added,
and the prelude setup has been tidied up.