Extend the exception interface and hierarchy
Exceptions now provide methods for retrieving information from their
backtraces, as well as access to the backtrace itself. The hierarchy of
errors has been split between the pre-existing Internal Error, which
represents a bug in the interpreter; Runtime Error, which indicates an
error in the runtime due to an erroneous program; and Logic Error, which
indicates an error as a result of a program that is semantically correct
but which has nonetheless produced an error. Two, unused classes of
exceptions now exist in the prelude, for future use, as well as a new
Logic Error that indicates use of an abstract method.

To facilitate this change, adds a fold() startingWith() method to
sequence (and therefore strings), and has made the tidy up of the
backtrace following use of assert() more consistent and reliable. A new
pattern called 'Foreign' has also been added, which matches against
JavaScript objects that are not Grace objects.