Retrieve dialect methods directly from the object
The existing dialect mechanism required that the object set to 'self' at
the top-most scope, but also that all the method names appeared in the
scope object. Now the searchScope method will look in self when it
reaches the top. This resolves a bug where custom dialects would not
respond to outer, as they did not have their methods copied into scope.