Ensure delegation sets the appropriate self value
The value of self in a Grace-defined method is always bound to the
surrounding scope, even if the actual method is called with a different
context. Now Grace methods will look at the context they were called in
and swap out the calling self if the two are different. This allows the
delegateTo() method to set the self value to the calling object.

Note that, because the value of self no longer necessarily reflects the
surrounding environment, a local search must look in both self and
scope, in the at order.