Add string utilities to String
The String implementation has had a number of utilities added to it,
including ordering, find and replace and index searching. The index
searching should also be on Sequence, but it is harder to implement due
to the JavaScript indexOf not allowing a custom comparator. The String
type has been updated to include these utilities.