Move setimmediate import to appropriate module
The import of the setimmediate module was in the interpreter rather than
in task, which produced an issue of import order if setImmediate was not
already present in the system. Now the task module will import
setimmediate before setting anything else up, ensuring that it exists
before the task mechanism is used.