Use foreign lookup when objects in scope
When an object was a Grace object, but had methods which were foreign,
there was no coercion of parameters and result, and no special lookup
for built-in methods occurred. The same applied to objects which were
not Grace objects, but appeared in scope. Now the lookup only considers
whether the requested method is a Grace method, and is applied to
lookups on a self or outer object in scope.