Remove single-part methods taking non-part arrays
Previously, passing the arguments of a single-part, non-generic method
call directly to the method in JavaScript was allowed as a special case
to permit a simple FFI system. Now this is the default, and passing a
JavaScript array that does not have a 'generics' property (which is how
part arguments are designated) to a single-part method will use the
array as a single argument, rather than its contents. Now passing normal
JavaScript arrays as parts is a special case for multi-part methods.

This commit also resolves an error in the Assertion Failure exception
that passed an array instead of a string to the string join method.