Remove the asPrimitiveString method on non-strings
The asPrimitiveString method was implemented on booleans and numbers as
a convenience. This confused the print() method, as it was identifying
these objects as strings, and not requesting the asString method. This
interfered with inheritance, where the asString method may be overridden
and return a different string representation than the one provided by
asPrimitiveString. Now only strings have the asPrimitiveString method.