Replace custom Unicode data with npm module
The existing unicode module was large and unwieldy, as well as somewhat
inefficient. Now the project dev-dependencies includes Unicode 7 data,
which provides regular expressions for matching character classes. This
is used at build-time to generate a unicode module. This unicode module
no longer exports name information, and the lexer just reports the
actual character in case of an unrecognised character. The 'escape'
function has been moved to the util module.

The asap dependency bounds have been extended and brfs has been moved to
a dev-dependency.