Use ESLint with custom rules instead of JSLint
ESLint is capable of checking more rules, and enforcing more stylistic
consistency. An ESLint configuration file is now included, which turns
on most optional rules, and the rest of the project has been modified to
conform to these rules. The major change is that array literals now
require padding. The consistent-this and and spaced-line-comment rules
require a minor patch to accept the existing style.