Produce better errors when failing to assign
Failing to find an assignment is a tricky error message, because it's
correct to report that a method "name :=" couldn't be found, but it's
less confusing to complain about the assignee. Now failed assignment
will check if the assignee exists, and report either that it cannot be
found or that it cannot be assigned to.