Correct special-case part descriptors for methods
Single-part descriptors for methods can be special-cased by supplying
the descriptor for just that part, but the detection method only
detected a number descriptor and not a 'gte' descriptor. Simultaneously,
the special-casing for inherited constructors was not performing any
sort of special-case detection. This caused the delegateTo() method to
fail in two distinct ways. Both of these use cases are now tested for,
and the special-case detection in both cases uses isArray() to fully
test for any special-cases.