Improve stringification of Request AST nodes
The RequestPart nodes in the AST now include generic parameters in their
stringification, and the rendering of operators is now more intelligent
about parameter use. Nested prefix operators and binary operators which
have a binary operator as a receiver will now correctly insert
parentheses, and binary operators will only insert parentheses on its
parameter when it is also a binary operator. It will still include
parentheses on the mathematical operators where precedence might permit
them to be omitted.