Replace type declarations with 'let' declarations
Let declarations are a more general form of 'static' declarations, which
permit things other than types from being defined in a non-overridable
way. This removes the awkward parsing following the 'type' keyword and
the special case in parsing for a single type literal. Type proxies are
now not available, but will be reintroduced when the bodies of type
literals are fully evaluated.

The full semantics of let declarations is still fairly undefined. The
hoisting seems fairly incompatible with permitting general expressions
to appear in the declaration, and it's still not entirely clear how the
forward-referencing in type bodies is going to work. This commit
represents just the generalisation of type declarations to let, and
further work remains on clarifying the behaviour of the new
declarations.