Correct the definition of & method for types
The & method on types used the 'self' variable before it was defined.
This fix moves the assignment to self to the top of the method (as a
best practice), as well using 'this' directly in the case where it was
broken. The combinatorial patterns now start with an upper-case letter,
following the standard for patterns.