Expose abstract and singleton pattern constructors
Implementations of patterns in Grace required reimplementing the whole
pattern interface. A 'pattern' class is now available in the prelude,
with an 'abstract' constructor which allows inheriting from the
AbstractPattern primitive.

A Singleton primitive has also been added, with a corresponding
'singleton' constructor in the pattern class, which generates a new
pattern that only matches itself. Note that this differs from the
singleton method in the runtime definitions, which produces a new
pattern that matches anything equal to the given object.