Add compareTo method to ordered objects
The Order type has been renamed to Ordered, and now requires that
implementing objects provide a compareTo method, which returns a
Comparison object. A Comparison object is either LessThan, EqualTo, or
GreaterThan (though like the Boolean type, there is nothing preventing
an object which is none or all of those values), and they provide an
interface for branching based on which of the three values they are.