Rename sequence to list, abstract array internals
The sequence class and Sequence type have been renamed to list and List
respectively, and the list class and List type have been renamed to
mutableList and MutableList. Both forms now stringify with the same
delimiters, and the custom asString method in mutable lists has been
removed. Lists now have first, last, and removeAt() methods.

The operations on an internal array of elements has been abstracted into
a separate constructor, in preparation for more collections. An
internalRemove() method has been introduced for collections which do not
have access to indexing, and lists now make use of this abstraction.