Add set and dictionary classes to the prelude
Mirroring the list and mutableList classes, sets and dictionaries are
now available in the prelude. The primitive Set and Dictionary
constructors inherit from InternalArray, so the collections are just
backed by a standard list and aren't particularly efficient.

The Collection type has been added, which abstracts some of the List
interface that the Set and Dictionary types also include. Strings now
have a contains() method in order to satisfy the Collection type. The
Set and Dictionary types are defined in the prelude, as they aren't used
internally in the interpreter. All collections now also have an
asImmutable method, as a mechanism for converting mutable collections to
their immutable form. The method is on immutable collections as well, in
order to accept collections which are either immutable or mutable and
immediately ensure that any possible mutable aliasing is removed.

All of the exposed definitions in the prelude are public, so that when
the prelude is delegated to they are still visible.