Extend the AST with more common types and visitors
All AST nodes now have an accept() method, which takes a visitor and
requests the appropriate method with itself, and passes it to accept()
on its own child nodes. A module for creating base visitors has been
added, and exported on the Node type along with the other node types.

Defs and Vars are now unified under a single Declaration type. The
'parameters' in a request part have been renamed to 'arguments', which
is a better representation of their purpose. Nodes with optional
components, like pattern annotations, have a patternOrIfAbsent() method
to allow their use in Grace, which cannot see null values.