Correct some erroneous behaviour in the primitives
Some methods in the primitive definitions assume that their arguments
are Grace objects, but they should be able to interact with JavaScript
objects. This requires using the apply function in the runtime instead
of directly calling the method on the object.

A few of the definitions in the string methods have some bugs fixed. The
upper index on substrings is up-to-not-including, and the startsWith()
and endsWith() methods return Grace booleans, not JavaScript booleans.