Add reflection through a mirrors module
The internal mirrors module defines a Mirror, which reflects an object,
as well as a MethodMirror, which is a first-class representation of a
method. The interpreter currently doesn't have a convenient way to
export an object as a module which can be imported through a path, so
the mirrors module is available in the prelude.