Import fs normally in the module loader
The fs module was being imported only on a call to the loader function
in order to handle it not appearing in the browser. The browserify
process permits it to appear so long as nothing is actually called on it
(with brfs intercepting the calls statically), so it is now imported as
usual. The brfs plugin still has trouble with applying readFile as a
task, so it is built into a Task in two stages.