Correct assert() code to handle other exceptions
Assigning to a variable before it is initialised is supposed to raise an
UndefinedValue exception, but it raised an InternalError exception
instead because the code for the assert() method on AbstractPattern did
not consider that a error might be raised as part of the assertion. If
this had worked, it would have been a different InternalError exception
about the match() method being undefined because of an incorrect call to
defs.pattern. This changes fixes both of these problems, and adds a test
that assigning to a variable before it is initialised fails in the
appropriate manner.