Update eslint, and update modules to match
The new version of eslint has some modified rules which allow for
greater customisation. As a result the array-bracket-spacing rule has
been set to the opposite of what it was before, necessitating a fairly
large change as every one-line array literal needs to be changed (note
that there may also be functions which could be pulled up onto the same
line now, but this has not been done). Several indentation mistakes have
also been identified and fixed.