Replace the custom tasks with native promises
In the interest of performance, the implementation of tasks is now
directly backed by native promises. This means that the custom
implementation of the 'now' and 'stop' methods are no longer supported.
The final use of the 'now' method, in the implementation of 'toString'
on all GraceObjects, has been replaced with a static string. The ability
to 'stop' has been moved into the interpreter, though this destroys the
interpreter's functionality. This should be tidied up in the future.

There is still some custom behaviour for tasks. Tasks are still run in a
context, which is preserved between calls to 'then'. This should be
removed in a future update when non-binding functions are used in 'then'
calls. Yielding to the event loop is also still implemented, and this
should be moved into the interpreter in the future as well.