Move event loop yield to the interpreter
In order to move custom behaviour out of the task implementation, the
interpreter is now responsible for yielding to the event loop. This also
has the benefit that the yield is not global: each interpreter is
responsible for yielding its own control. In general this functionality
is pretty shaky, and should be tidied up to give more control to the
client.