Remove tasks in favour of native promises
This commit removes the remainder of the custom functionality of task
objects, and replaces them entirely with native promises. This means
that asynchronous code no longer preserves the context the code is
executing in. This has far reaching implications for most modules, but
in general is fairly solvable by replacing the regular functions with
the ES2015 arrow function syntax, causing the context to be lexically
bound. A lot of these transformations have been made automatically.

The task module still exists, but just exports the static functionality
that is still relevant to promises.